feat: useCart 훅을 추가하여 장바구니 관리 로직 분리
